[xmerl] Fix XML "well-formedness" bug i SAX parser
Changed the XML Sax parser behaviour so it returns an error
if there are something more in the file after the the
matching document. If one using the xmerl_sax_parser:stream()
a rest is allowed which then can be sent to a new call of
xmerl_sax_parser:stream() to parse next document.

This is done to be compliant with XML conformance tests.