zip: Don't crash when a zip file has an Unix extra header
There is unfinished code in the zip module to handle the
extra field in the central-directory header for each
file. The code to extract an Unix extra subfield is incorrect
and always causes a bad_unix_extra_field error. A further
flaw in the code is that it expects that there is only
a single subfield.

Correcting the code for extracting extra subfields will not do any
good, because the extracted data is ignored. In fact, not even the
file times extracted from DOS file times are used for some reason.

Therefore, don't correct the unfinished code. Remove it completely.

(If needed, extending zip to use file times and to use the information
in the extra field should be done in a major release, not in a
maintenance release.)

ERL-348