dialyzer, observer: Use portable lookup of home directory
Use init:get_argument(home) to find the location of the home
directory. That will work on all platforms (including Windows).

Note that the run-time system will fail to start if HOME
(or the equivalent on Windows) is not set. Therefore, it can
be assumed that init:get_argument(home) will not fail.

ERL-161