Add transport_opt() capx_strictness
To allow the Peer State Machine requirement that only the expected
capabilities exchange message be received in the relevant state to be
relaxed. If {capx_strictness, false} is configured then anything bu the
expected CER/CEA is ignored.

This is non-standard behaviour, and thusfar undocumented. Use at your
own risk.