Travis CI: Avoid that builds gets killed for inactivity
Travis CI will kill build jobs that have not produced output for
10 minutes.

The OTP build was never killed because of inactivity before 74796de9c7739
(which started to capture the output in a temporary file). After that
commit, now and then a build would be killed because it did not finish
in 10 minutes.

Update the build script to periodically print a ".", but only if the
size of the log file has changed. That way, if there is a real hanging
during the build, Travis CI will still kill the build.

Alternatives considered: Using travis_wait. Rejected because travis_wait
will extend the build unconditonally, even if the build does not progress
at all.