stdlib: Improve handling of pids, ports, and refs in qlc
The extended parser introduced in last commit is used in qlc for
solving an old bug: pids and refs could not be parsed by
string_to_handle(). The parser is also used for adjustments regarding
ETS identifiers (now references) in qlc_SUITE.

Notice that pids, references, ports, and external functions that
cannot be created in the currently running system cause syntax errors
(as before).