build-docker-otp: Avoid copying the git repository to docker daemon
Since docker is run from the top-level of the otp repository, the
entire repository will be the context for the docker daemon. In my
private repository, more than 800Mb had to be copied before the
build could be started.

Let the ./scripts directory be the context for docker instead.
That way, only the otp.tar.gz file needs to be copied.

While we are it, also use the ADD command in the Dockerfiles to
pack up the tar file and avoid copying the tar file itself to
the Docker image.