[ct] Add 'keep_logs' option
If setting the value for this option to an integer, N, common_test
will remove all ct_run.* directories in the current log directory,
except the N newest.

The default value for the 'keep_logs' option is 'all', which means that
no logs will be deleted.

'keep_logs' can be used in combination with refresh_logs, or in a normal
common_test test run.