Smoke test HiPE in the 32-bit build
HiPE is tested when running dialyzer in the 64-bit build. To avoid
running out of memory or time, dialyzer is not run in the 32-bit
build job.

Do a smoke test of HiPE by letting dialyzer create a small PLT.

We could catch more bugs by using 'configure --enable-native-libs',
but I am worried that the build would not finish in 50 minutes
(the time limit for a Travis job).