Add cross-compile of HiPE-enabled VM
To build a HiPE-enabled VM, a binary called hipe_mkliterals for the same
architecture that will host the VM must be run in order to generate header files
with structure offsets and other compiler- and architecture-specific information
HiPE needs at compile time.

In order to cross-compile a HiPE-enabled VM, the build process needs a way of
running hipe_mkliterals. The configure script now accepts a variable RUN which
is supposed to be a tool that accepts a binary for the VM architecture and
arguments, and will produce the same standard output that the binary would if
run on those arguments.

The configure script will try to autodetect QEMU, which is one possible choice
for RUN, as well as the dynamic linking path prefix for QEMU.