Partially revert commit 37d63e9b8a0a96125ada858205286a58a5bed1ae
The above commit erroneously declared 'ppc64le' as a 'ppc64' for
HiPE enabling. These two architectures are different and one can
NOT expect that native code generated for one architecture runs
on the other one.  Rather than reverting this commit, make sure
that the architecture is declared correctly instead of pretending
it's a PPC64.