Rework gen_sctp suite to demonstrate remaining problems
In particular, that transmission can be very slow. The problem appears
to be linked to sndbuf/recbuf, but even with buffers that are large
enough to hold all messages being sent, turnaround times can still vary
by hundreds of milliseconds in a reasonable test environment.

Use multiple streams and a sender process to more closely mirror the
usage in diameter_sctp, but neither is the source of the problems.