Correct compile_SUITE:core_roundtrip/1
The test is supposed to compare the Core Erlang code that has been
printed and parsed back. It did compare and print any differences,
but it did not fail when there were differences.

Also fix problems with variable names and maps not comparing
equal when the inliner has been used.