stdlib: Add Unicode modifier t to control sequences w and W
As of the introduction of Unicode characters in atoms, the control
sequences 'w' and 'W' can return non-Latin-1 characters, unless some
measure is taken.

This commit makes sure that '~w' and '~W' always return Latin-1
characters, or bytes, which can be output to ports or written to raw
files.

The Unicode translation modifier 't' is needed to return non-Latin-1
characters.