Restore undocumented Failed-AVP setting convenience
The parent commit removed the convenience of setting something like the
following in the errors field of the diameter_packet of an answer
message.

  [#diameter_avp{} = A2, {5001, #diameter_avp{} = A1}]

This results in Result-Code = 5001 and Failed-AVP = [A1,A2], but is
currently undocumented. Probably useful, so restore it.

Also accept {RC, [#diameter_avp{}]} at encode, which is probably more
useful; eg. [{5001, [A || {5001, A} <- Errors]}]

Anyone who wants full control can set errors = false and formulate
Result-Code/Failed-AVP themselves. (As opposed to not setting a value
explicitly, which results in setting from the decoded errors list. A bit
quirky, but documented and historical.)