Mend leak of stacktrace in catch and try...catch
p->ftrace holds the stacktrace from the last exception in
the process. If the last exception was a 'function_clause' or
a 'badarg' from a BIF call, the arguments for the call are also
stored in p->ftrace. The arguments could potentially be huge.

p->ftrace is user by erlang:get_stacktrace/0 to retrieve the
stacktrace for last exception in a process.

To eliminate this leak, we must do a slightly incomptible change
to how erlang:get_stacktrace/0 works. It will only return a stacktrace
when called from within the scope of try...catch. The following will
work:

  try
    ...
  catch
    C:R ->
      Stk = erlang:get_stacktrace(), %% Will return a stacktrace.
      {C,R,Stk}
  end.

The following will also work:

  try
    ...
  catch
    C:R ->
      Stk = helper(),
      {C,R,Stk}
  end.

helper() ->
  erlang:get_stacktrace().  %% Will return a stacktrace.

The following will no longer work:

  try
    ...
  catch
    C:R ->
      ...
  end,
  erlang:get_stacktrace() %% Will return [].

There is a another complication. Consider this code:

  try
    ...
  catch
    C:R ->
      tail_recursive_helper(C, R)
  end.

tail_recursive_helper(C, R) ->
  {C,R,erlang:get_stacktrace()}.

If we insert an instruction that clears the stacktrace after the call
to tail_recursive_helper/2, the call will no longer be tail-recursive.
If we want to preserve tail recursion and still make
erlang:get_stacktrace/0 return a stacktrace, we can't clear the
stacktrace in p->ftrace.

Since destroying tail recursion could have worse consequences than
keeping the stacktrace, we will not clear the stacktrace in this case.

In the case of an old-style 'catch', erlang:get_stacktrace/0 will work
if it is used before the function that executes the 'catch' returns.
That is, the following examples will work:

f1() ->
  ...,
  (catch ...),
  ...,
  erlang:get_stacktrace(),
  ...

f2() ->
  ...,
  (catch ...),
  ...,
  {error,helper()}.

helper() ->
  erlang:get_stacktrace().

If there is a tail-recursive call to a helper function, the
stacktrace will not be cleared from the process.

f3() ->
  ...,
  (catch ...),
  ...,
  helper().

helper() ->
  erlang:get_stacktrace().