crypto: use lazy evaluation for test cases read from CAVP files
It turns out that the excessive memory usage is cause by the
test framework printing all the test vectors into the log output.

A similar proplem was already diagnosed for long_msg/0. The root
cause was not mentioned in the SUITE, but the same fix applies
to the CAPV test vector data.

Switch all CAPV data to lazy evaluation and have the test itself
read the data.